Skills, Armor, and Weapons
Added Starting Skills, Updated Weapon and Armor details

+++ src/equipment/Item.java
@@ -2,5 +2,5 @@ package equipment;
 
 public interface Item {
 
-  public double getWeight();
+  public int getWeight();
 }
